requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.23
botocore==1.29.23
certifi==2022.9.24
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
dj-database-url==1.0.0
Django==4.1.3
django-environ==0.9.0
django-rest-framework==0.1.0
django-s3-storage==0.13.11
django-storages==1.13.1
djangorestframework==3.14.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
numpy==1.23.5
pandas==1.5.2
placebo==0.9.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.2.0
urllib3==1.26.13
Werkzeug==2.2.2
zappa==0.56.1
